fix: look for schema tag in extensions
